requirements.txt
joblib==1.2.0
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.1
streamlit==1.17.0
